Deprecate a few bond constructors.
The idea is to have a single, most used constructor for each
kind of bond and to use the generic constructors in the Bond
class for special cases.  This makes it a lot more convenient
to use the constructors in languages (like Python or C#) that
support some kind of syntax for keyword arguments.