CPICoupon with baseDate
add a new CPICoupon constructor which requires a baseDate instead of base CPI fixing value. If baseCPI is provided it will use it, if it baseCPI is NULL the pricer will retriev the index fixing at baseDate from the index.

If baseDate and baseCPI are null and we have at least two dates in the schedule, CPILag will imply a regular baseDate = firstCouponStartDate - observationLag
If the CPILeg has only one date and one cashflow, a non-null baseDate or baseCPI are still required