Allow creating PiecewiseZeroInflationCurve before baseDate is known
QuantLib's yield curves can be created before fetching market data,
which allows creating complex object graphs ahead of time and setting
market data (quotes and fixings) at a later stage. We use this feature
extensively in our code.

The new API for PiecewiseZeroInflationCurve does not offer the same
flexibility, since it requires baseDate to be set at the curve creation
time. BaseDate comes from inflation fixings, which means that the
fixings have to be fetched before the curve can be created.

Add another constructor for PiecewiseZeroInflationCurve, which takes a
function that returns the baseDate. This function will be called during
the curve's bootstrap, and the result will be cached. This allows one to
construct PiecewiseZeroInflationCurve first and fetch the fixings at a
later stage.