Add test showing issue with deferred observer.
The test gives a read access violation. The problem is that some
elements in deferredObservers_ can be destroyed while
ObservableSettings::enableUpdates() is running. Later in the call
deferredObserver->update() is called on the already destructed element
and it gives a read access violation.

The example here uses the zero coupon inflation swap helper to
demonstrate the issue. The zciis_ member is one of the deferred
observers. When ObservableSettings::enableUpdates() is running
ZeroCouponInflationSwapHelper::initializeDates() is called because the
helper is notified and updateDates_ is true. zciis_ is re-assigned and
the original pointee is destroyed. But the original pointee that has
been destroyed is still in the deferredObservers_ set and update() gets
called on it leading to the read access violation.